Srw/cert seed app

+++ CCIA/Models/StateProvince.cs
@@ -7,7 +7,7 @@ namespace CCIA.Models
     {
         public int StateProvinceId { get; set; }
         public string StateProvinceCode { get; set; }
-        public short CountryId { get; set; }
+        public int CountryId { get; set; }
         public string StateProvinceName { get; set; }
         public DateTime DateModified { get; set; }
